fix friends list
