+++ study-connect/app/profile/[uid]/page.tsx
@@ -3,11 +3,9 @@ import { useState, useEffect, Suspense } from "react";
 import { useRouter, useParams } from "next/navigation";
 import dynamic from "next/dynamic";
 import { db, auth } from '../../../lib/firebase';
-import { doc, getDoc, updateDoc, arrayUnion, setDoc, collection, query, where, getDocs } from 'firebase/firestore';
+import { doc, getDoc, updateDoc, arrayUnion } from 'firebase/firestore';
 import { User, FriendRequest } from '../../utils/interfaces';
 import { v4 as uuidv4 } from 'uuid';
-import Image from 'next/image';
-import Link from 'next/link';
 
 const ProfileContent = dynamic(() => import('../../components/ProfileContent'), {
     ssr: false,
@@ -23,8 +21,6 @@ export default function UserProfile() {
     const [currentUser, setCurrentUser] = useState<User | null>(null);
     const [friendStatus, setFriendStatus] = useState<'none' | 'friends' | 'pending' | 'requested'>('none');
     const [isLoading, setIsLoading] = useState(false);
-    const [friendsList, setFriendsList] = useState<User[]>([]);
-    const [loadingFriends, setLoadingFriends] = useState(false);
     const { uid } = useParams(); 
 
     useEffect(() => {
@@ -83,50 +79,6 @@ export default function UserProfile() {
 
         fetchUserProfile();
     }, [uid]);
-
-    // Fetch friends data when user is loaded
-    useEffect(() => {
-        if (!user || !user.friends || user.friends.length === 0) {
-            setFriendsList([]);
-            return;
-        }
-        
-        const fetchFriends = async () => {
-            setLoadingFriends(true);
-            try {
-                const friendsData: User[] = [];
-                
-                // Fetch each friend's data
-                for (const friendId of user.friends) {
-                    const friendDoc = await getDoc(doc(db, 'users', friendId));
-                    if (friendDoc.exists()) {
-                        const data = friendDoc.data();
-                        friendsData.push({
-                            userId: friendId,
-                            name: data.name || '',
-                            email: data.email || '',
-                            grade: data.grade || '',
-                            major: data.major || '',
-                            minor: data.minor || '',
-                            joinedClasses: data.joinedClasses || [],
-                            profilePic: data.profilePic || 'https://upload.wikimedia.org/wikipedia/commons/a/ac/Default_pfp.jpg',
-                            aboutMe: data.aboutMe || '',
-                            friends: data.friends || [],
-                            friendRequests: data.friendRequests || [],
-                        });
-                    }
-                }
-                
-                setFriendsList(friendsData);
-            } catch (error) {
-                console.error("Error fetching friends: ", error);
-            } finally {
-                setLoadingFriends(false);
-            }
-        };
-        
-        fetchFriends();
-    }, [user]);
 
     // Check friend status when both users are loaded
     useEffect(() => {
@@ -236,50 +188,6 @@ export default function UserProfile() {
         }
     };
 
-    const renderFriendsList = () => {
-        if (loadingFriends) {
-            return (
-                <div className="flex justify-center py-8">
-                    <div className="animate-spin rounded-full h-6 w-6 border-b-2 border-blue-500"></div>
-                </div>
-            );
-        }
-
-        if (friendsList.length === 0) {
-            return (
-                <div className="text-center py-8 text-gray-500">
-                    {user?.userId === currentUser?.userId 
-                        ? "You don't have any friends yet. Add some friends to connect!"
-                        : `${user?.name} doesn't have any friends yet.`}
-                </div>
-            );
-        }
-
-        return (
-            <div className="grid grid-cols-1 md:grid-cols-2 lg:grid-cols-3 gap-4">
-                {friendsList.map(friend => (
-                    <Link 
-                        href={`/profile/${friend.userId}`} 
-                        key={friend.userId}
-                        className="flex items-center p-4 bg-white rounded-lg shadow-sm hover:shadow-md transition-shadow"
-                    >
-                        <Image 
-                            src={friend.profilePic} 
-                            alt={friend.name}
-                            width={50}
-                            height={50}
-                            className="rounded-full mr-4"
-                        />
-                        <div>
-                            <h3 className="font-medium text-gray-900">{friend.name}</h3>
-                            <p className="text-sm text-gray-500">{friend.major}</p>
-                        </div>
-                    </Link>
-                ))}
-            </div>
-        );
-    };
-
     if (!user) {
         return (
             <div className="h-screen flex items-center justify-center">
@@ -294,12 +202,6 @@ export default function UserProfile() {
                 {renderFriendButton()}
             </div>
             <ProfileContent user={user} setUser={setUser} />
-            
-            {/* Friends List Section */}
-            <div className="mt-8 px-8">
-                <h2 className="text-2xl font-bold mb-4">Friends ({user.friends.length})</h2>
-                {renderFriendsList()}
-            </div>
         </Suspense>
     );
 }
