+++ study-connect/next.config.ts
@@ -2,6 +2,9 @@ import type { NextConfig } from "next";
 
 const nextConfig: NextConfig = {
   /* config options here */
+  images: {
+    domains: ['upload.wikimedia.org', 'res.cloudinary.com'],
+  },
 };
 
 export default nextConfig;
