+++ team/retrospectives/RETRO_03.md
@@ -1,0 +1,26 @@
+# Date of Retrospective: 03/05/25
+
+* Led by: [Shelly Zhu]
+* Present: Allen Hu, Anthony Jin (Remote), Maria Saucedo-Flores, Shelly Zhu, Wesley Chiba, Zhenyu Yu, Hannah Su
+* Absent: None
+
+## Action item
+
+* Goal: Each member of the team should keep everyone up to date on what they are currently working on using the Kanban board as we work towards putting everything together for a more functional and complete app.
+
+* Change/Experiment: Keep the Kanban board updated with what we are working on. Everyone should have a task at all times, and let people know if the task is taking longer than expected.
+
+* Measurement: Everyone should be assigned to one issue on the Kanban board at all times. If an issue is completed by someone and the PR is merged, then that person should be assigned to another task. There is communication on Slack is someone takes longer than expected on a task.
+
+
+## Retro Assessment 
+
+* We used the Start, Stop, Continue retro format. We had the team type up what they thought we should start, stop, continue in the team Slack and submit at the same time. Then we had people share what they wrote and elaborate further, and then came up with a goal based on what people said we should start doing the most.
+* I think the retro went well. We seemed to all think that the communication on Slack was good and everyone was able to elaborate on what was going well and what we needed to work on.
+* I think the next person can try a different retro format to see if that brings out more ideas or makes it easier for people to speak up. They can also try starting the retro at the start of class and concluding it at the end so people have more time to think.
+
+## Experiment/Change
+
+* Our experiment was for everyone to keep the team updated on what was being worked on using the Kanban board to make more progress towards a more functional and complete app. I think we made good progress initially upon declaring the experiment and updating the Kanban board to have more issues and have issues for more people. But after we finished those issues, we did not immediately follow up with new ones so the experiment did not work as well as we had hoped it would.
+* However, we did make more progress between the time of starting the experiment and the time of this assessment, so I believe that updating the issues on the Kanban board at the start did help realign everyone's goals and what issues needed to be tackled. We also kept each other up to date on Slack on what was being worked on regularly and if tasks took longer than expected. People also did start another task after finishing the one they were currently working on, but it was not always reflected on the Kanban board.
+* Overall, I think the experiment did have a positive effect and we made more progress due to it, even if the Kanban board part of the experiment was not perfectly executed. Going forward, we will strive to keep the Kanban board updated along with continuing our regular updates and progress towards our complete app.
